requirements.txt
python-dotenv==0.21.0
requests==2.28.1
twilio==7.14.1
telegram
